Small fixes

+++ linux/install.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-echo "Installing Tek's environment....."
+RED='\033[0;31m'
+NC='\033[0m' # No Color
+
+echo "$RED Installing Tek's environment.....$NC"
 
 #Install OH MY ZSH
-echo "Installing OH MY ZSH"
+echo "$RED Installing OH MY ZSH $NC"
 sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
 sleep 1 
-echo "Installing OH MY ZSH Plugins"
+echo "$RED Installing OH MY ZSH Plugins $NC"
 
 #Install plugins
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
@@ -16,7 +19,7 @@ sleep 1
 
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 
-echo "Installing P10K"
+echo "$RED Installing P10K $NC"
 
 #Install theme
 git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
@@ -27,19 +30,28 @@ sleep 1
 useConfirm=true
 
 confirm() {
-   [ "$useConfirm" = true ] && read -p "Proceed? (Enter) - (^C to abort)"
+   [ "$useConfirm" = true ] && read -p "$NC Proceed? (Enter) - $RED (^C to abort) $NC"
 }
 
-echo "Copying settings" 
+echo "$RED Copying settings $NC" 
 confirm
 
 current_time=$(date "+%Y.%m.%d-%H.%M.%S")
 echo "Current Time : $current_time"
 
-mkdir ~/.bak
-mv -v ~/.zshrc ~/.bak/zshrc.$current_time
+mkdir ~/.bak -v
+
+[ -f ~/.zshrc ] && mv -v ~/.zshrc ~/.bak/zshrc.$current_time
 cp -v .zshrc ~/.zshrc
-mv -v ~/.alias ~/.bak/alias.$current_time
+
+[ -f ~/.alias ] && mv -v ~/.alias ~/.bak/alias.$current_time
 cp -v .alias ~/.alias
-mv -v ~/.p10k.zsh ~/.bak/p10k.zsh.$current_time
+
+[ -f ~/.p1-k] && mv -v ~/.p10k.zsh ~/.bak/p10k.zsh.$current_time
 cp -v .p10k.zsh ~/.p10k.zsh
+
+[ -f ~/.dir_colors ] && mv -v ~/.alias ~/.bak/dir_colors.$current_time
+cp -v .dir_colors ~/.dir_colors
+
+#[ -f ~/.dir_colors ] && mv -v ~/.alias ~/.bak/dir_colors.$current_time
+#cp -v .dir_colors ~/.dir_colors
